Move TabCounter to a separate module
