+++ app/build.gradle
@@ -194,6 +194,7 @@ dependencies {
 
     implementation project(':firebase')
     implementation project(':feature-tabs')
+    implementation project(':ui-tabcounter')
 
     implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk8:${Versions.kotlin}"
