+++ ui-tabcounter/src/main/res/drawable/tab_counter_box.xml
