+++ ui-tabcounter/src/main/res/values/attrs.xml
@@ -1,0 +1,11 @@
+<?xml version="1.0" encoding="utf-8"?>
+<!-- This Source Code Form is subject to the terms of the Mozilla Public
+   - License, v. 2.0. If a copy of the MPL was not distributed with this
+   - file, You can obtain one at http://mozilla.org/MPL/2.0/. -->
+<resources>
+
+    <declare-styleable name="TabCounter">
+        <attr name="drawableColor" format="color" />
+    </declare-styleable>
+
+</resources>
