+++ ui-tabcounter/src/main/res/values/colors.xml
@@ -1,0 +1,5 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources>
+    <color name="colorMenuIconForeground">@color/sharedColorAppPaletteBlack</color>
+    <color name="sharedColorAppPaletteBlack">#FF272727</color>
+</resources>
